download_libs: allow to not overwrite for multi downloads

+++ scripts/dev/download_libs.sh
@@ -3,6 +3,12 @@ set -e
 VER=$1
 PLATFORM=$2
 ARCH=$3
+if [ $# -gt 4 ]; then
+    OVERWRITE=$4
+else
+    OVERWRITE=1
+fi
+
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 if [ "$PLATFORM" == "msys2" ] || [ "$PLATFORM" == "vs" ]; then
@@ -17,19 +23,29 @@ wget ci.openframeworks.cc/libs/$PKG
 cd ../../
 mkdir -p libs
 cd libs
-echo "Removing old libraries"
-libs=("boost" "cairo" "curl" "FreeImage" "freetype" "glew" "glfw" "json" "libpng" "openssl" "pixman" "poco" "rtAudio" "tess2" "uriparser" "utf8" "videoInput" "zlib" "README.md")
-for lib in $libs; do
-    if [ -e $lib ]; then
-        rm -rf $lib
-    fi
-done
 
-echo "Uncompressing libraries"
-if [ "$PLATFORM" == "msys2" ] || [ "$PLATFORM" == "vs" ]; then
-    unzip -qo ../scripts/dev/$PKG
+if [ $OVERWRITE -eq 1 ]; then
+    echo "Removing old libraries"
+    libs=("boost" "cairo" "curl" "FreeImage" "freetype" "glew" "glfw" "json" "libpng" "openssl" "pixman" "poco" "rtAudio" "tess2" "uriparser" "utf8" "videoInput" "zlib" "README.md")
+    for lib in $libs; do
+        if [ -e $lib ]; then
+            rm -rf $lib
+        fi
+    done
+    echo "Uncompressing libraries $PLATFORM $ARCH"
+    if [ "$PLATFORM" == "msys2" ] || [ "$PLATFORM" == "vs" ]; then
+        unzip -qo ../scripts/dev/$PKG
+    else
+        tar xjf ../scripts/dev/$PKG
+    fi
 else
-    tar xjf ../scripts/dev/$PKG
+    echo "Uncompressing libraries $PLATFORM $ARCH"
+    if [ "$PLATFORM" == "msys2" ] || [ "$PLATFORM" == "vs" ]; then
+        unzip -qn ../scripts/dev/$PKG
+    else
+        tar xjkf ../scripts/dev/$PKG
+    fi
 fi
+
 
 rm ../scripts/dev/$PKG
